Makes Chrome extension more dev-friendly
The extension now checks for the presense of a specific div, and only if it
exists will it modify the page. This way, we can have the extension run on
localhost and ngrok domains in addition to the joinoneroom domain, without the
fear of the extension modifying someone else's page.